==> notebooks/1.0-data-loading.ipynb <==
### Objectives of this notebook:  
1. Unzip the ParaNMT-detox dataset to get around GitHub's limit on upload file size.
2. Perform basic data understanding of the dataset.

# Unzipping the dataset in the raw directory:
import zipfile

with zipfile.ZipFile("../data/raw/filtered_paranmt.zip", mode="r") as archive:
    archive.printdir()

with zipfile.ZipFile("../data/raw/filtered_paranmt.zip", mode="r") as archive:
    dataset = archive.read("filtered.tsv").decode(encoding="utf-8")
    with open("../data/interim/filtered_paranmt.tsv", "w", encoding="utf-8") as f:
        f.write(dataset)


Now that I have unzipped the dataset, I can get acquianted with it with the help of pandas library:

import pandas as pd

dataset = pd.read_csv("../data/interim/filtered_paranmt.tsv", delimiter='\t')
dataset = dataset.set_index(dataset.columns[0])
dataset.index.name = "Index"

dataset.head()

dataset = dataset.rename(columns={'lenght_diff': 'length_diff'})
dataset.info()

All the values in the dataset are already non-null.

dataset.dtypes

We can infer that the dataset shape is (577777, 6).

dataset.describe()

### Conclusion:  

Dataset contains 2 categorical columns: sentence to be detoxified and its paraphrised version.  
It contains 4 numerical columns: cosine similarity of texts, relative length difference between texts, toxicity level of the original text and toxicity level of its paraphrased version.  
Dataset is already cleaned from null values.  
Observing the toxicity values, I notice that the std of both ref_tox and trn_tox are similar, but differing mean suggests that messages are considered to be more toxic before their paraphrasing.  
Looking at the similarity value, I see that the mean is 0.758469, so the necessary conclusion is that a lot of paraphrasing loses important meaning carried by the original text, which could cause the disparity in toxicity values.
