==> notebooks/2.0-detox-preparation.ipynb <==
We will be attempting to solve detoxification using models from "Text Detoxification using Large Pre-trained Neural Models" by David Dale et al. Let us first clone their repository.

!git clone https://github.com/s-nlp/detox ../detox

libraries = []
with open('../detox/requirements.txt', 'r') as f:
    libraries = f.read().splitlines()

libraries

Check to see if there are any packages not installed.

%pip install -r ../detox/requirements.txt -q

### Dataset Preparation  
As we remember, some of the translations in our dataset have a higher toxicity score than the original statements. Keeping that in mind, let us form a new dataset, that would reliably contain toxic and non-toxic sentences in different columns.

import pandas as pd

dataset = pd.read_csv("../data/interim/filtered_paranmt.tsv", delimiter='\t')
dataset = dataset.set_index(dataset.columns[0])
dataset.index.name = "Index"

# Create new columns based on aforementioned conditions
dataset['toxic'] = dataset.apply(lambda row: row['translation'] if row['ref_tox'] < row['trn_tox'] else row['reference'], axis=1)
dataset['non-toxic'] = dataset.apply(lambda row: row['reference'] if row['ref_tox'] < row['trn_tox'] else row['translation'], axis=1)
dataset['old_toxicity'] = dataset.apply(lambda row: row['trn_tox'] if row['ref_tox'] < row['trn_tox'] else row['ref_tox'], axis=1)
dataset['new_toxicity'] = dataset.apply(lambda row: row['ref_tox'] if row['ref_tox'] < row['trn_tox'] else row['trn_tox'], axis=1)
dataset['toxic'] = dataset['toxic'].str.lower()
dataset['non-toxic'] = dataset['non-toxic'].str.lower()
dataset.drop(['reference', 'translation', 'similarity', 'lenght_diff', 'ref_tox', 'trn_tox'], axis=1, inplace=True)
dataset.head()

dataset.describe()

Now the toxic and non-toxic sentences are clearly separated in the dataset, so we have something to compare the performance of our dataset against.

dataset.to_csv("../data/interim/separated_tox.csv")

### Considering Metrics  

For evaluation purposes we will be using metrics from the aforementioned paper: *J*-score, *ACC*, which were implemented in the PMLDL_Assignment1\detox\emnlp2021\metric\metric.py file. The comparison with the existing dataset will be conducted through the ACC metric, as we cannot afford to manually gather the data on the toxicity of reformulated sentences by the model from many different people.
